requirements.txt
pillow
requests
pytz
slack-sdk
